Fix paging half-heartedly
This adds paging to the history/ section so we're not displaying *all*
the results.

The paging kind of sucks though since we have to pull back all the results
from redis since they're stored by keys and not by timestamp and we want
to sort them by timestamp before paging.