wrap string returns to browser with bytes()
This is needed for PEP3333 support, which is required for
gunicorn-0.17.2 and above. Note that in reality, only strings using
external variables (namely '(who, app_name, webapp_ref)') need this. For
bare strings with no (or only local) variables seem to work fine. For
consistency, though, we just do everything.